修复新闻封面显示错误问题

+++ lib/core/models/user.dart
@@ -7,7 +7,7 @@ class UserMo {
   String? sex;
   String? avatar;
   String? idCard;
-  int? balance;
+  double? balance;
   int? score;
 
   UserMo({
